requirements.txt
asgiref==3.4.0
boto3==1.17.111
botocore==1.20.111
dj-database-url==0.5.0
Django==3.1.12
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-storages==1.11.1
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.0
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
